moved enums to a separate class
