--- app/src/main/java/com/example/happymeals/AmountUnit.java
@@ -1,5 +1,0 @@
-package com.example.happymeals;
-
-public enum AmountUnit {
-    count, mg, mL
-}
