+++ app/src/main/java/com/example/happymeals/Constant.java
@@ -1,0 +1,14 @@
+package com.example.happymeals;
+
+public class Constant {
+    public enum Location {
+        Pantry, Fridge, Freezer
+    }
+    public enum IngredientCategory {
+        Dairy, Meat, Fruit, Vegetable, Snack
+    }
+    public enum AmountUnit {
+        count, mg, mL
+    }
+
+}
