--- app/src/main/java/com/example/happymeals/IngredientCategory.java
@@ -1,5 +1,0 @@
-package com.example.happymeals;
-
-public enum IngredientCategory {
-    Dairy, Meat, Fruit, Vegetable, Snack
-}
