+++ app/src/main/java/com/example/happymeals/IngredientStorageActivity.java
@@ -20,7 +20,7 @@ public class IngredientStorageActivity extends AppCompatActivity {
         // TODO: pull from database
         // remove from here
         storage.setIngredients();
-        storage.addIngredient(new Ingredient("Apple", "01-11-2022", Location.Fridge, 1, AmountUnit.count, IngredientCategory.Fruit));
+        storage.addIngredient(new Ingredient("Apple", "01-11-2022", Constant.Location.Fridge, 1, Constant.AmountUnit.count, Constant.IngredientCategory.Fruit));
         // remove to here
 
         storageListView = findViewById(R.id.storage_list);
