--- app/src/main/java/com/example/happymeals/Location.java
@@ -1,5 +1,0 @@
-package com.example.happymeals;
-
-public enum Location {
-    Pantry, Fridge, Freezer
-}
